requirements.txt
bitcoinrpc-auth-proxy>=1.0.0
python-dotenv>=0.15.0
python-bitcoinlib>=0.11.0
pillow>=9.0.0
logging